feat(gitbrowse): allow custom branch (#172)
## Description
Allow the user to be able to define custom branch with
`Snacks.gitbrowse`. I also added `start_line` and `end_line` (default to
`nil`), because from my testing when using a keymap such as the one in
[my comment in the related
issue](https://github.com/folke/snacks.nvim/issues/170#issuecomment-2508476110)
or this one which I came up later while implementing this
```lua
vim.keymap.set({ "n", "x" }, "<leader>gY", function()
  local function system(cmd, err)
    local proc = vim.fn.system(cmd)
    if vim.v.shell_error ~= 0 then
      Snacks.notify.error({ err, proc }, { title = "Git Browse" })
      error(err)
    end
    return vim.split(vim.trim(proc), "\n")
  end
  local branches = {}
  local start_line, end_line
  if vim.fn.mode() == "v" or vim.fn.mode() == "V" then
    start_line = vim.fn.line("v")
    end_line = vim.fn.line(".")
    if start_line > end_line then
      start_line, end_line = end_line, start_line
    end
  end
  for _, branch in ipairs(system({ "git", "branch" }, "Failed to get git branches")) do
    branch = branch:match("%s*[%*]?%s*(%S+)")
    table.insert(branches, branch)
  end
  vim.ui.select(branches, {
    prompt = "Select branch",
  }, function(choice)
    if choice then
      Snacks.gitbrowse({ start_line = start_line or nil, end_line = end_line or nil, branch = choice })
    end
  end)
end, { desc = "Git Browse ()" })
```
Then the visual lines would not work. My guess is that in the first case
the command line prompt and in the second case the `vim.ui.select`
before calling `Snacks.gitbrowse` break the Visual mode state and any
visual selection is not propagated to `Snacks.gitbrowse` to be properly
evaluated by the internal code for visual lines selection, which works
just fine when you just call directly `Snacks.gitbrowse()` via a
function in a keymap (which I assume doesn't break Visual mode state).
That's why I had to implement the visual selection again in the keymap
as well and needed the extra parameters to pass the newly evaluated
values.

Please do correct me if I'm somehow wrong in my logical deduction (which
is based mostly on observation and guessing) and maybe `start_line` and
`end_line` are not needed. Except if the user passes the option directly
in the keymap, but then he would have to create multiple keymaps for
different branches, which in my personal opinion defeats the purpose of
this feature.

In both cases of the keymaps I mentioned the user can dynamically either
type or choose the branch available to him. But maybe I'm missing
something in both the keymaps implementations that I came up with.
<!-- Describe the big picture of your changes to communicate to the
maintainers
  why we should accept this pull request. -->

## Related Issue(s)
Closes #170.
<!--
  If this PR fixes any issues, please link to the issue here.
  - Fixes #<issue_number>
-->

## Screenshots

<!-- Add screenshots of the changes if applicable. -->

---------

Co-authored-by: Folke Lemaitre <folke.lemaitre@gmail.com>