fix(bigfile): check existence of NoMatchParen before executing (#561)
## Description

<!-- Describe the big picture of your changes to communicate to the
maintainers
  why we should accept this pull request. -->

In my config I use the following flag to disable the default
`MatchParen` plugin.

```
vim.g.loaded_matchparen = 1
```

This causes an error when `bigfile` tries to call the `NoMatchParen`
command since it does not exist.
This PR fixes that by checking for the existence of the command before
executing it.

## Related Issue(s)

<!--
  If this PR fixes any issues, please link to the issue here.
  - Fixes #<issue_number>
-->

## Screenshots

<!-- Add screenshots of the changes if applicable. -->