fix(zen): sync cursor with parent window (#547)
## Description

The cursor position is now synchronized between the zen/zoom window and
it's parent.

I'm not sure if this is the best way to achieve that. Initally I wanted
to do set the parent's cursor position only when the zen buffer is
closed, but I couldn't figure out a way to do that.

## Related Issue(s)

Fixes #539