fix(picker.icons): fix typo in Nerd Fonts and display the full category name (#716)
## Description

Thanks for the icons picker 🙇🏼 

I had one in my dotfiles I was going to contribute, but you have been
faster.
So there it is, I added some things I had in mine

This PR:
- fix a typo in the Nerd Fonts source name (its plural)
- add The category full names to be consistent with emojis
 
<!-- Describe the big picture of your changes to communicate to the
maintainers
  why we should accept this pull request. -->

## Related Issue(s)

<!--
  If this PR fixes any issues, please link to the issue here.
  - Fixes #<issue_number>
-->

## Screenshots

![image](https://github.com/user-attachments/assets/0691b3ac-665a-4d53-92ad-aa3cc56d1546)

<!-- Add screenshots of the changes if applicable. -->

> [!NOTE]
> Would you be interested in some extra sources I was working on ? At
least Math and Box drawing ?

> [!NOTE]
> I would also switch the name the source so Source can be capitalize
like the Category and I think having Source > Category after is easier
to read as the glyph and the icons are actuelly what people search. But
this is personal taste, sp I did not submnitted this change