fix(filter): insert path from `filter.paths` into `self.paths` (#761)
## Description
Here we see that we insert the paths into `filter` instead of
`self.paths`
https://github.com/folke/snacks.nvim/blob/b96bd540f785c725289f9f15f0147b1b2dac5a35/lua/snacks/picker/core/filter.lua#L34-L44
and later on here
https://github.com/folke/snacks.nvim/blob/b96bd540f785c725289f9f15f0147b1b2dac5a35/lua/snacks/picker/core/filter.lua#L90-L94
we check `self.paths` to get the path.

This was causing issues to not correctly filter out the already defined
paths and I was able to see, for example the scratch buffer files in the
pickers.
<!-- Describe the big picture of your changes to communicate to the
maintainers
  why we should accept this pull request. -->

## Related Issue(s)
None
<!--
  If this PR fixes any issues, please link to the issue here.
  - Fixes #<issue_number>
-->

## Screenshots

<!-- Add screenshots of the changes if applicable. -->