feat(scope): allow injected languages to be parsed by treesitter (#823)
## Description

Adding a config flag to allow injected languages to be parsed by
treesitter. This is crucial for languages like vue, which contain
typescript or javascript in a script tag. When `treesitter.get_node` is
not called with ignore_injections = false, the script tag is treated as
a single block, which is not helpful, as it contains most code for
languages like vue.

## Related Issue(s)

 - Fixes #430 

## Screenshots

 - See #430

---------

Co-authored-by: Folke Lemaitre <folke.lemaitre@gmail.com>