feat(image): add support for `svelte` (#1277)
## Description

Added queries for Svelte as well, based on Vue.

Note that I am not sure if the current HTML queries, which both Vue and
this Svelt query inherits from, is working as intended.

I do not get any sort of inline images. But I guess that will have to be
debugged and fixed in another issue and the fix should not affect this
patch (hopefully).

## Related Issue(s)

Fixes #1275

---------

Co-authored-by: Folke Lemaitre <folke.lemaitre@gmail.com>